backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.71",
    "fastapi>=0.135.1",
    "mangum>=0.21.0",
    "openai>=2.26.0",
    "pypdf>=6.9.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
